dev/auto_stock/requirements.txt
asgiref==3.10.0
django==4.2.26
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
requests==2.32.3
python-dotenv==1.2.1
